[RSCH-10] Run Elixir provision tools
Problem: Development of web-apps with Elixir is complex.

Solution:
In order to address the complexity, Elixir community came up with `mix` tool
which, among other things, allows the user to provision web-applications that
use Phoenix web framework.

Included in this commit are the results of running
  + `mix phx.new`, which generates a Phoenix-compatible Elixir project
  + `mix ecto.create`, which creates the databases as specified in `config/`

Ommitted in this commit is a reproducible way to initialize PostgreSQL
database. However, we have provided the listing to do that at home in `README.md`.